Finish separating Qt from Python code.
